Started using Open3D.
I want to integrate further parts of the pipeline in the C++
application. This require the app to start using Open3D, which is
better anyway, so that we can stop generating raw dumps of the point
clouds and directly load the RGBD images.

The classes could probably have a better architecture, but I wanted to
get to this point, so I preferred drafting these mechanisms for now.