Button to restore the original matrix in align.
Sometimes the alignment produces bad/unexpected results, so having a
button to cancel the last alignment is useful.

In the future, I might improve this, and write the matrix in the scene
only when accepting the result, but at the moment we do not have a way
to specify another matrix when rendering.