Add a way to edit raw matrices.
When using a raw matrix, now we show it.
However, dealing with matrices is not easy, therefore I added a way to
chain more transformations.
I find it very convenient to use, therefore I might remove the other way
to transform point clouds.

Also, this UI is useful to transform multiple matrices at once (useful,
e.g., for when you aligned a few frames, and you want to transform them
to be closer to another set of aligned frames, or if you need to adjust
many frames for symmetry purposes).