Implemented RGBD masks.
Sometimes, depth images have errors, especially close to where the data
becomes invalid. This might result in a lot of noise in those areas.
If you are using just a few frames, it might be feasible to mark some
areas as not to be used for the TSDF (merge) phase.

This commit allows adding a .png for each frame in a directory named
"mask". All pixels whose alpha value is less than 128 will be considered
as invalid (equivalent to depth = 0).