Add a feature to symmetrize the frames.
Being able to export a more or less centered mesh should be helpful to
use the result (e.g., to re-mesh only half of it) even though not
necesarily realistic.

The symmetrize function tries to divide the pointcloud in the positive X
and negative X half-spaces. Then, it mirrors one of the two, and runs
ICP on it. In this way, we find a transformation that we should apply
to make the two half overlap, and we apply half of it the point cloud.
We do this iteratively, to possibly find the best transformation.

I do not have any analysis on the convergence of the rotation to 0
(actually, with real world data at a certain point the detected angle
started oscillating), however empirically this works well enough for me.

Once accepted, the new transformation will be applied to all the frames
that were used to create the merged point cloud. However, we run it on
the already merged one to possibly work on more organized data.