Apply texture to a (processed) mesh (PoC).
The "Texture lab" allows to load a triangle mesh that matches more or
less the point cloud data (e.g., it can be the raw output of the mesh,
or its refinement made with tools such as Instant Meshes) and applies
the RGB data as texture.

Currently, it is just a proof of concept, because it just shows the
texture applied. I still have not decided how to continue the
development of this tool.