Show the mesh with multiple textures.
We now find the various vertices in all the selected frames.
We also allow to export the mesh with these textures applied.
We still don't have a UI to choose which texture to use in each
triangle, for now we go in frame order.

This commit doesn't work with the current main of Open3D for me, because
of a bug in the KDTreeFlann implementation.
I opened a PR (isl-org/Open3D PR #6734) to fix this bug.
A local build of Open3D with the fix applied works correctly for me.