Updated dependencies, CMakelists.txt, and README.md.
This commits cleans a little bit the third-party dependencies.

We start including glm as a submodule, wgucg might simplify the effort
to get started with this project.

Also, move all the CMake definitions related to the submodules  (and to
glad) to a CMakeLists.txt on their own.

Now, we rely only on existing dependencies that are also Open3D
dependencies. I added that info to the README.md.
Since I was at it, I clarified also other points.