Created an application to work on models with bones.
Rigid registration is somehow limited, and cannot work around slight
variations, e.g., on limbs.
I thought that maybe skinning the mesh built with TSDF might be a
possibility for precise alignments.

This commit adds whatever needed to create the bones structure
interactively, but it does not actually render a skinned mesh, and it
does not include a functionality to set bone weights.