Added ImGuizmo support and copy transformation.
This commit makes it possible to use ImGuizmo to transform point clouds
both when editing a single frame and in multi-edit mode.
The UX could probably be improved (e.g., allow to choose the Gizmo
reference for rotations and support the transformation stack).

Also, added a button to copy the tranformation of another frameset.