Allow using scalable volumes.
I initially used only uniform volumes because they allow to set a limit
to the volume and the memory they need is easily predictable.
However, depending on the dataset, this might not be a problem, and
scalable dataset have often better performances in my experience.
So, since the changes to support them were minimal, the merge
functionality now lets choose between uniform or scalable.