Support PNG color images in the GUI.
In this commit, AddFrameState has been modified to match color and depth
images according only to their stem (filename without extension).
In this way, we support both JPEG and PNG images, and we also are case
insensitive on the file extension.

As a result, a depth image might match more than one color image.
Our choice was to take only the first one we find (which depends on the
directory traversal order).

The directories are still hardcoded to rgb and depth, maybe that will
change in the future.

A side effect of this change is that we use narrow char strings also on
Windows, so we can use strnatcmp also there (even though this program
has never been built on Windows, so there might be other problems).